87% complete

+++ pixel.h
@@ -1,6 +1,7 @@
 #ifndef PIXEL_H
 #define PIXEL_H
 typedef struct {
-    int red, green, blue;
+    float red, green, blue ;
+
 } Pixel;
 #endif // PIXEL_H
